Make example style less confusing
Both title and heading1 override fontName, so the fontName in heading
doesn't have any effect. Move this to title and heading1 instead where
they are effective.

Make the 200% fontSize in title explicit.

These changes don't have any effect on the output, but make the style
file easier to follow and change.